This is my second revision
I have a circle which will eventually be my diver following my mouse
nicely using void mouseMoved(). The only issue I have with that part is
to make sure once the ellipse is moved it does not disappear. That will
be in my next revision. I also added the x & y coordinates of the mouse