Third Revision
I have made my sketch much more efficient. I realized with the help of
Mr. Gordon that if the code was in voidMousemove that the ellipse would
only be drawn when the mouse was moved. I instead moved it back into
voiddraw used the integers already given to us by processing known as
mouseX and mouse Y and now my code is much more efficient.