I have a diver!
i took an image off of the internet just for the diver and I
implemented it as my diver. I also now have an awesome bubble system in
which 5 bubbles are released from the divers helmet, they all travel at
different speeds and once they touch the boundaries they regenerate in
the divers helmet. The next thing I need to do is make a working
collision course that works with this diver. Lastly I have to create
sharks instead of the squares.